requirements.txt
python-telegram-bot>=20,<22
httpx>=0.27.0,<0.28
beautifulsoup4>=4.12.0,<5
lxml>=5.0.0,<6
python-dotenv>=1.0.0,<2
